% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{raadsync}
\alias{raadsync}
\alias{raadsync-package}
\title{\pkg{raadsync}}
\description{
\pkg{raadsync}
}
\author{
Australian Antarctic Data Centre \email{aadc@aad.gov.au}
}
\references{
\url{http://www1.data.antarctica.gov.au}
}

