% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_batch.R
\name{sync_batch}
\alias{sync_batch}
\title{Synchronise a raadsync repo using BatchJobs}
\usage{
sync_batch(config, reg, create_root = FALSE, verbose = TRUE)
}
\arguments{
\item{config}{data.frame: raadsync configuration as returned by \code{\link{read_repo_config}}}

\item{reg}{Registry: a BatchJobs registry object}

\item{create_root}{logical: should the data root directory be created if it does not exist?}

\item{verbose}{logical: if TRUE, provide additional progress output}
}
\value{
list, with status and captured console output from each synchronisation task
}
\description{
Note that the \code{reg} object should be created with raadsync in the list of packages (see example).
}
\examples{
\dontrun{
 library(BatchJobs)
 setConfig(cluster.functions= makeClusterFunctionsMulticore(ncpus=7),debug=FALSE)
 reg <- makeRegistry(id="syncysync",file.dir="sync_files",packages="raadsync")
 cf <- read_repo_config("/my/local/repo/config")
 sync_log <- sync_batch(cf,reg)
 removeRegistry(reg,ask="no") ## clean up
}

}

