% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sync_repo}
\alias{sync_repo}
\title{Run the data repository synchronisation}
\usage{
sync_repo(config, create_root = FALSE, verbose = TRUE)
}
\arguments{
\item{config}{data.frame: configuration as returned by \link{\code{read_repo_config}}}

\item{create_root}{logical: should the data root directory be created if it does not exist?}

\item{verbose}{logical: if TRUE, provide additional progress output}
}
\description{
Run the data repository synchronisation
}

