% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadsync.R
\docType{package}
\name{raadsync}
\alias{raadsync}
\alias{raadsync-package}
\title{\pkg{raadsync}}
\description{
\pkg{raadsync}
}
\author{
Australian Antarctic Data Centre \email{aadc@aad.gov.au}
}
\references{
\url{http://.data.aad.gov.au}
}

