% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_repo_config.R
\name{read_repo_config}
\alias{read_repo_config}
\title{Load data repository configuration file}
\usage{
read_repo_config(local_config_file,
  default_config_file = system.file("extdata", "raad_repo_config.json",
  package = "raadsync"))
}
\arguments{
\item{local_config_file}{string: file or URL to JSON configuration file.}

\item{default_config_file}{string: file or URL to JSON configuration file. By default, the default_config_file is the one that ships with this package.}
}
\description{
This configuration file specifies global settings that control the synchronization behaviour in general, and provides details of each of the datasets in the repository. Any global configuration setting can also be set at the dataset level. A dataset inherits any global configuration settings not overridden in its own configuration. It is suggested that a user only makes changes to the local_config_file, because package updates may change the default configuration file provided with the package. To use only a local_config_file, specify NULL for the default_config_file (and vice-versa).
}
\details{
The global settings include the following parameters:
\itemize{
\item wget_flags string: flags to pass to wget (default is "--no-passive-ftp")
\item http_proxy string: http proxy server to use in the form "http://your.proxy.server:8080" (default is no proxy server)
\item ftp_proxy string: ftp proxy server to use in the form "ftp://your.ftp.proxy.server:21" (default is no proxy server)
\item local_file_root string: the path to the data root directory. If the \code{raadtools} library has been installed it defines a default data directory option. This will be used if no local_file_root is supplied in the \code{raadsync} global configuration
\item clobber numeric: 0=do not overwrite existing files, 1=overwrite existing files if the remote copy is newer than the local copy, 2=overwrite existing files unconditionally
\item wait numeric: seconds to wait between downloads (default=0)
}

Each dataset is specified by the following parameters in the configuration file:
\itemize{
\item name string: dataset name
\item description string: description
\item reference string: URL to metadata record or home page for the dataset
\item source_urls list: list of URLs to the data
\item do_sync logical: if FALSE this dataset will not be synchronized
\item method string: the synchronization method (currently only "wget")
\item method_flags string: flags to pass to wget (if method=="wget") or the custom handler (if not)
\item postprocess string array: operations to apply after all file downloads. Currently "unzip", "gunzip" (unzip or gunzip but do not delete the compressed copy), "gunzip_delete", "unzip_delete" (unzip and delete the zipped file), "cleanup <pattern>" (remove any files in local_directory with names matching <pattern>), "cleanup_recursive <pattern>" (as for cleanup but is applied recursively to child directories)
\item user string: username for access to restricted ftp/http sites
\item password string: password for access to restricted ftp/http sites
}
}

