% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sync_repo.R
\name{sync_repo}
\alias{sync_repo}
\title{Run the data repository synchronisation}
\usage{
sync_repo(config, create_root = FALSE, verbose = TRUE)
}
\arguments{
\item{config}{data.frame: configuration as returned by \code{\link{read_repo_config}}}

\item{create_root}{logical: should the data root directory be created if it does not exist?}

\item{verbose}{logical: if TRUE, provide additional progress output}
}
\description{
Run the data repository synchronisation
}

